Initial Commit - send `context_type` frame attribute
Rails tracebacks are loooong. Displaying the entire traceback in sentry makes me
sad, so lets augment the data we are sending by using Rails' backtrace_cleaner
to specify which frames are 'application' and which are 'framework'. (This is
using the same mechanism that the rails 500 page uses to display the filtered
tracebacks).

Change-Id: I8880f7f4aa5002112977c6d82ba4ec31ae8cd1ad