Send `cleaned_function` attributes for backtrace frames
Rails frequently includes frames with gross dynamic function names. One
of the most frequent offenders are callbacks, which take the form:

    _run__4437211773795971890__process_action__1264216456344873512__callbacks

This is breaking sentry's exception grouping. Let's send a
'cleaned_function' attribute in which we replace all integer substrings
with the constant 'N'. This may be problematic for applications in which
there are functions whose names differ only by an integer; that seems
like a rare enough edge case to ignore for now. If it becomes
problematic later we can write a more clever replacement algorithm.